requirements.txt
Keras==2.4.3
opencv_python==4.5.1.48
tensorflow==2.4.1
numpy==1.19.5
streamlit==0.80.0
Pillow==8.2.0